add HTMLAttributes to SvgIcon - move types - ad key to iconPaths
