+++ react-svg-bubble-slider/src/_docs/2.icons.stories.mdx
@@ -16,7 +16,7 @@ To provide a restricted array of icons for your SvgBubbleSlider see <Link href='
     flexWrap: 'wrap',
   }}
 >
-  {iconPaths.map((icon, index) => {
+  {iconPaths.chrisGannon.map((icon, index) => {
     const { name, path } = icon
     return (
       <Flex
